Move files to correct directory and format with Black
