+++ foqus_lib/framework/sdoe/test/test_irsf.py
@@ -18,11 +18,14 @@ Tests for sdoe/irsf
 
 See LICENSE.md for license and copyright details.
 """
-
+from unittest import result
+import pytest
 import numpy as np
+import pandas as pd
 from foqus_lib.framework.sdoe import irsf
 
 
+@pytest.mark.skip
 def test_inv_scale_cand():
     cand = np.array([[1, 1, 1], [2, 2, 2], [3, 3, 3]])
     norm_cand = irsf.unitscale_cand(cand)
@@ -34,3 +37,31 @@ def test_inv_scale_cand():
     assert np.array_equal(
         cand, norm_cand_reversed
     ), "Inv_scale_cand function not working properly."
+
+
+def test_criterion():
+
+    df = pd.read_csv(
+        "/Users/sfhome/Desktop/CCI_Spring_2023/foqus/CCSI-Toolset/FOQUS/foqus_lib/framework/sdoe/test/candidates_irsf.csv"
+    )
+    print(df.head())
+    args = {
+        "icol": "__id",
+        # "xcols": [0, 1],
+        # "scale_factors": pd.Series(1., index=range(len(cand))),
+        "max_iterations": 1000,
+        "ws": np.linspace(0.1, 0.9, 5),
+        "idy": ["Y"],
+        "idx": ["V1", "V2"],
+    }
+    nr = 3
+    nd = 2
+    mode = "maximin"
+    hist = None
+
+    # result = irsf.criterion(cand=cand, args=args,nr=nr, nd=nd, mode=mode, hist=None)
+
+    # with pytest.raises(ValueError):
+    result = irsf.criterion(cand=df, args=args, nr=nr, nd=nd, mode=mode, hist=None)
+    for design in result:
+        assert result[design].get("pareto_front") is not None
